Add a very basic initial testbench for Esch384
The testbench is currently set up to run 10 times to provide an average.
Each run consists of running both the reference and optimised implementation
for every length of message up to MAX_MESSAGE_LENGTH (currently 4096),
and timing the result for both implementations. In the future it may be
better to instead have a set message length, and time 2000 iterations of
that, for instance. This would also allow measuring whether longer/shorter
messages are affected more by the optimisations or not.

The reference implementation used here comes from the NIST submission at:
https://csrc.nist.gov/CSRC/media/Projects/lightweight-cryptography/documents/finalist-round/updated-submissions/sparkle.zip
Specifically, the subfolder '[zip]\Implementations\crypto_hash\esch384v2\ref'.
The SIMD-optimised implementation currently is just a copy of the reference
implementation, however this will be the basis for the optimised code.