Fix up the project structure, and start adding Schwaemm256_256 to the testbench
The compilation of each algorithm is now gated behind defines in
src/toggle_defines.h. This avoids function naming collisions, though
this can be fixed another way too. The Sparkle permutation is the same
for both Esch and Schwaemm (so far), so has been split into it's own
folder and is used by both algorithms. The optimised version may need
different optimisations depending on the parameters, but that will be
determined later.

Schwaemm currently runs in the testbench and outputs all the test
vectors, validated against those in the NIST submission. This will be
expanded soon to comparing the SIMD version against the reference
version, as with Esch384. Timing also needs to be added for that.