Add an initial SIMD-optimised version of Sparkle
This initial implementation for the SIMD optimisations focuses on
the underlying Sparkle permutation. In the reference algorithm,
this runs a number of iterations of the ARXBOX layer for each step,
depending on the number of branches (4, 6, and 8 for Sparkle256,
Sparkle384, and Sparkle512 respectively). In the SIMD implementation,
each step instead runs all iterations of the ARXBOX layer in parallel
by splitting the state into two 256-bit vector registers. Each
registers contains half the state, and the ARXBOX operations are
performed using C intrinsics, before the state is finally stored
back to be used in the linear layer.
This is almost certainly not the optimal implementation as of yet,
so the next step will be to optimise it to increase performance.

Currently, it appears to run approximately 10% faster in the best
case (Esch384 with 1<<16 bytes as input). This performance increase
will likely be lower for the algorithms using Sparkle256 and 384,
as these can only take up 4/8ths or 6/8ths of the vector registers.