Overhaul testbench Esch timing, and test some SIMD optimisations
The timing method for Esch has been overhauled to take multiple measurements
and calculate an average. The same will need to be done with Schwaemm as well.

Additionally, I've explored the performance differences with a different
order for the indices in the gather instructions. The current implementation
gathers the state using indices that are not separated by a set stride.
Experimenting with using indices that are consecutive with a stride of 2 (e.g.
0, 2, 4, 6... etc) in fact yielded worse performance, likely due to it then
being necessary to add an addition permute4x64 instruction for the hi and lo
state prior to the unpacking. It may however still be worth looking in to
further, as loading the state fully consecutively (i.e. 0, 1, 2... etc) may still
prove to be faster, if an efficient way of shuffling can be found.