Add initial NEON implementation, and various other fixes
The initial implementation using Arm NEON instructions has now
been added. Defines have been added to separate NEON/AVX2 code,
and to split platform-specific code (e.g. Windows/Linux timing).

The NEON implementation is fairly similar to AVX2, except that
the registers are only 128 bits compared to 256 bits with AVX2.
This means that for the fully 512-bit sparkle state, 4 SIMD
registers are needed instead of the 2 for AVX2. This is split
into 2 pairs, with one pair containing the first 8 state elements,
and the second containing the last 8. The operations on each
pair are exactly the same, just the 4 alzette rounds of rotation,
addition, and XOR. Storing is fairly efficient, requiring only
a single zip and 8-element memcpy per pair. Cases have been
added for sparkle384 and sparkle256, to avoid overrunning when
writing back to state.

Additionally, the ref decryption function was being used in the
testbench for Schwaemm, so now correctly uses the simd function.
The compiler options for CMake were also not working before, as
they must be specified *before* add_executable. This has been
fixed.