Finish testbench timing for the Sparkle Suite
All algorithms now have proper timing, with the testbench timing
functions reworked to make the process easier. Of note, the
averages now exclude any result with a difference more than 50%
in order to remove outliers, as often a single run would be much
slower/faster and throw off the average.
The results currently look like as follows on the Raspberry Pi:
- Schwaemm128_128: ~8% faster than ref
- Schwaemm192_192: ~9% slower than ref
- Schwaemm256_128: ~8% slower than ref
- Schwaemm256_256: ~6% faster than ref
- Esch256: ~10% slower than ref
- Esch384: ~7% faster than ref

The current conclusion from these results is that algorithms
using the Sparkle256/512 permutations run noticeably faster
than the reference algorithms, while those using Sparkle384
run noticeably slower. My theory for the cause of this is that
in the case of Sparkle384, Alzette still requires 2 runs even
though the second run only half fills the SIMD registers.
For Sparkle256, only 1 full run is required, and Sparkle512
has 2 full runs.
This would suggest that the performance gain from the second
run does not outweigh the overhead incurred from the SIMD
instructions as only 2 pairs are being processed in parallel.
It may be more performant to do 1 SIMD run, and process the
remaining two pairs using the original reference algorithm.