More optimisation to the NEON Sparkle implementation
For Sparkle384, process only the first 4 pairs using SIMD.
Processing the last 2 pairs with SIMD is inefficient as the
NEON registers are only half full. Instead, just use the
reference Alzette steps. This brings the performance of the
algorithms using Sparkle384 (Esch256, Schwaemm192192 etc)
about on par with the reference algorithm. This is an improvement
from them being ~8% slower previously, but it may still not
be worth using SIMD at all for that permutation variant.
Whether this will be the same with AVX2 remains to be seen,
and will be investigated next.