Experiment with unrolling the linear layer for Sparkle
I had the idea to try keeping the state in the SIMD registers when
running the linear layer, with the idea that avoiding the extra
load/stores could be benficial for performance. This proves to be
counterproductive, as accessing individual elements in the SIMD
registers through a union can actually force the data to be flushed
to memory first. This causes the performance to be much worse than
reference (~30% slower), and so is not worth it.