Add Pyjamask128 algorithm to the testbench, and improve outlier removal
The Pyjamask128 AEAD algorithm will be the next one that I look at for this
project, as it appears to be a good candidate for SIMD optimisation. It was
a round 2 submission to the NIST lightweight cryptography standardisation
process, so it is similar to SPARKLE in that respect. The NIST-submitted
algorithms all appear to expose a similar AEAD interface (I believe this is
so that they can be used in the SUPERCOP benchmark), so the Schwaemm timing
and verification methods have been repurposed as general AEAD methods that
work for both Schwaemm and Pyjamask.

Additionally, outlier handling has been reworked. Instead of removing result
over 50% from reference, the standardized Z-scores are calculated for the
dataset. Any result more than 2 standard deviations from the mean is removed,
and the average is recalculated to account for this. This should be a more
accurate way of minimizing the outliers in the results, especially when more
results are taken for the final report. I will also look into methods such
as core pinning and locking clock speeds to try and minimize variation overall.