[SPARKLE] Some code formatting for the report, and bug fixes
In multiple places, the instruction _mm256_xor_epi32 was used. This
intrinsic can compile to AVX512 instructions or AVX2 instructions
depending on the optimisation level. In these cases it was pure luck
that it was compiling to an AVX2 version, but for the in-progress
Pyjamask implementation, this compiled to an AVX512 instruction and
subsequently caused issues. Instead, use the AVX2 _mm256_xor_si256
intrinsic. The above also applies to the use of _mm_xor_epi32 for
the 128-bit implementation.