[PYJAMASK] Implement an AVX2 8-block parallel version of Pyjamask128
This commit contains the initial AVX2 implementation of Pyjamask128.
As the algorithm uses the OCB mode of operation, it is possible to
encrypt multiple blocks in parallel. With 256-bit YMM registers,
8 128-bit blocks are encrypted/decrypted at once. This is only
possible when there is at least 8 fully blocks in the input, and in
cases with less blocks, the serial algorithm runs as normal. Any
remaining blocks not in a group of 8 are also processed serially as
before.

To run 8 blocks in parallel, we first have to perform all the
per-block preprocessing in serial, such as calculating the offset,
L_i etc. This has to be done in order since each iteration relies
on the previous calculation. Once this is done however, the actual
block cipher can be run for all 8 blocks at once. After, the result
can be copied back to the output as normal.

Parallel 'x8' versions of all the necessary block cipher methods have
been added, which operate on 256-bit vector registers instead of 32-bit
integers. This is largely a 1-1 translation of the serial version. The
only important differences are in the load/unload state methods. Instead
of packing the 8-bit input values into 32-bit integers and then gathering
those, we instead gather them first and shuffle the order of bytes
within each register afterwards. Whether this is more efficient should
be tested, but a single shuffle per vector register seems like it
should be faster than manually packing every 8-bit integer first. For
unloading, the reverse is done. The lack of scatter instructions makes
it harder to store the state back, so it is largely done by just
serially copying values back to memory.