[PYJAMASK] Fix loop for processing 8 blocks at once
This maths was completely off, and was essentially causing only 1/8th
of the total blocks to be computed using the parallel algorithm.
For example with 4096 blocks as input, only the first 512 would be
split into chunks of 8 and computed in parallel. The last 3584 blocks
would all be computed serially.

This massively improves performance with a lot of blocks, bringing the
timings up from only ~8% faster than reference to ~62% faster than
reference. These results are more in line with what I was expecting to
see with the degree of parallelism achieved here.