[PYJAMASK] Optimise matrix multiplication function
The pj128simd_mat_mult_x8 function was taking the largest amount of time in
profiling due to shifting a YMM register taking 3 SIMD instructions. It is
much quicker to just keep mat_col as a uint32_t, and broadcast it to a YMM
register when necessary. Performance improvement goes from +65% to +73%.