[TESTBENCH] Improve methodology for timing
There were a few issues remaining with the previous timing method that have
been fixed here:

Printed output was too messy:
The data arrays are no longer printed, and the data output has been minimized.
This can be restored by passing the debug flag to time_aead/time_esch.

Average time taken was not reported, only average difference:
Instead of calculating the difference for each pair of results, we now calculate
the average for both the reference and SIMD implementations. A % difference is
still calculated, but reporting the time average allows for proper graph plotting
for the report.

Outlier removal was flawed:
Although we already applied a single pass of Z-score normalization to the data,
I found that it was not sufficient when the data contained some extreme outliers.
In that case, only the extreme outliers were removed, while outliers that were
not so extreme were still kept. Now, we perform 2 passes of outlier removal, which
does a much better job of keeping the results consistent. In my testing, no more
than about 10 values are removed out of 50 when we have the highest variance,
with it being a lot less most of the time.

These changes, combined with raising the number of runs to 50, gives much more
consistent and accurate results. This will make reporting timings for the
report much easier, as the results are much more repeatable.

Additionally, I have begun work to support running multiple times with different
parameters. For the report, it will be necessary to gather timings with various
values for the input and AD lengths, so the end goal will be to dump the raw
results to a file from which they can be used in e.g. Matplotlib to produce graphs.