[PYJAMASK] Add Pyjamask96 to the testbench, and implement AVX2 version
The Pyjamask96 algorithm has been added to the testbench. In the same way as with
Sparkle, the underlying primitive for both Pyjamask96 and Pyjamask128 - the Pyjamask
block cipher - has been added to a separate folder and shared between both algorithms.

The AVX2 implementation is largely trivial, just following the same method as with
the AVX2 implementation of Pyjamask128. We still run 8 blocks in parallel, but each
block is now only 96 bits and so only 3 256-bit state registers are needed. The only
big changes are to the functions to load and unload state, which have been modified to
work with a state_size of both 3 and 4.

Preliminary results show an improvement of about 72-73%, which is nearly exactly the
same as with Pyjamask96. The total time taken is slightly higher though, which is
likely because we have to perform more block cipher invocations even if each is
slightly smaller.