[PYJAMASK] Add full Neon implementations for Pyjamask128/96
This commit adds full implementations of Pyjamask96/128 using ARM
Neon instructions. Overall it is mostly a 1-1 translation of the
AVX2 instructions to their 128-bit Neon equivalents, although in
some cases the Neon version uses different instructions such as
bitwise NOT on the vector instead of XOR-ing with 1's. Instead
of running 8 blocks in parallel, we can only run 4. The methods
in encrypt.c have been changed to do this for Neon, kept separate
from the AVX2 8-block loop to keep things simple.

Since Neon has no 4-vector zip/unzip, state is just manually
loaded/unloaded with packing like in the serial implementation. This
may not be the optimal implementation, but performance is still very
good.

Performance is approx. 60% improvement over reference for both
encryption and decryption in both algorithms. This is slightly less
than the 8-block version, though not as much less as I thought it
may be.