Improve platform checks to support non-MSVC platforms with AVX2
The CMake configuration has been updated to allow running the AVX2
implementation on platforms such as Linux, without being locked into
MSVC. This has allowed benchmarking and testing on GCC with WSL2, which
shows a marked difference in performance. It also highlighted somewhere
that _mm_or_epi32 was used, which is in fact an AVX512 instruction. This
has been replaced with the SSE2 equivalent, _mm_or_si128.