Fix outlier removal crash, and change float printing in results
While testing results with lower input lengths, a crash started happening in
clean_and_calculate results. With a lower input length, the result times were
low enough (e.g. 20us) that often the an outlier pass would leave a results
array with all identical values. This caused issues with standard deviation
calculation as the sum was 0, so in this case we just skip this iteration since
there are no outliers to remove.

Additionally, only print results to 1 d.p for the report.